+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- move asset received hooks to handler so it can check protocol whitelist
 - add some extra parse utils used when debugging invariant tests
 - whitelist erc721/1155 in invariant tests and ensure all assets are actually being received by teller from swapper
 - deposit invariant tests add special case workaround for retrieved gas comp == gas comp cap (see inline commment)